third_party/WebKit/LayoutTests/inspector/domdebugger/domdebugger-getEventListeners-expected.txt

Tests RemoteObject.eventListeners.


Running: testSetUp

Running: testWindowEventListeners
type: scroll
useCapture: true
location: 11:16
handler: function foo() {}
sourceURL: domdebugger-getEventListeners.html


Running: testDivEventListeners
type: click
useCapture: true
location: 12:16
handler: function boo() {}
sourceURL: domdebugger-getEventListeners.html

type: click
useCapture: false
location: 8:37
handler: function onclick(event) {
  return 42;
}
sourceURL: domdebugger-getEventListeners.html

type: mouseout
useCapture: false
location: 11:16
handler: function foo() {}
sourceURL: domdebugger-getEventListeners.html


Running: testDivWithoutEventListeners
